feat(documentation): add comprehensive source of truth document
This commit introduces a detailed "Source of Truth" document for the project, aimed at serving as a master reference for building resumes, preparing for interviews, and providing a quick-reference guide. The document outlines the service's purpose, architecture, tech stack, and operational procedures, ensuring clarity and consistency for current and future team members. This addition enhances knowledge sharing and onboarding processes, reducing the learning curve for new developers.

+++ services/transaction-processor/src/main/java/com/tradestream/transaction_processor/api/TransactionDto.java
@@ -13,7 +13,7 @@ public record TransactionDto(
         UUID userId,
         String side,
         String ticker,
-        int quantity,
+        BigDecimal quantity,   // <—
         BigDecimal price,
         Instant executedAt
 ) {
